google-storage-bucket-source-5g7t/requirements.txt
quixstreams[s3]==3.7.0
python-dotenv
tempfile
google-cloud-storage
